requirements.txt
PyQt5>=5.15.6
Pillow>=9.0.0
UnityPy>=1.9.0
numpy>=1.20.0
texture2ddecoder>=1.0.3 